Add getSetByCode() and split tests
